refactor: ResultStatus를 판단하는 로직을 ResultStatus로 이동
