+++ src/main/kotlin/blackjack/domain/Gambler.kt
@@ -18,17 +18,6 @@ class Gambler(name: String) : Participant(name) {
     }
 
     fun determineResult(dealer: Dealer): GamblerResult {
-        return GamblerResult(this, determineResultStatus(dealer))
-    }
-
-    private fun determineResultStatus(dealer: Dealer): ResultStatus {
-        return when {
-            dealer.isBlackjack() && this.isNotBlackjack() -> ResultStatus.DEFEAT
-            dealer.isBurst() -> ResultStatus.WIN
-            this.isBurst() -> ResultStatus.DEFEAT
-            this.isScoreLargerThan(dealer) -> ResultStatus.WIN
-            this.isScoreEqualTo(dealer) -> ResultStatus.DRAW
-            else -> ResultStatus.DEFEAT
-        }
+        return GamblerResult(this, ResultStatus.of(this, dealer))
     }
 }
